% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cancel_receipt}
\alias{cancel_receipt}
\title{Cancel an emergency message}
\usage{
cancel_receipt(receipt, ...)
}
\arguments{
\item{receipt}{A message receipt (e.g., 'KAWXTswy4cekx6vZbHBKbCKk1c1fdf')}

\item{...}{An application token can be specified with \code{token}}
}
\value{
A \code{\link{PushoverResponse}} object containing the response from
the server
}
\description{
\code{cancel_receipt} issues a query to Pushover to cancel sending the
emergency message with the given receipt.
}
\note{
Message receipts can be acquired using \code{\link{receipt}} on the
\code{\link{PushoverResponse}} object storing the response to an emergency
message.

The token argument is necessary, however it does not need to be given
if the application token have been set with \code{\link{set_pushover_app}}.
}

