% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valid_app.R
\name{check_valid_app}
\alias{check_valid_app}
\alias{test_valid_app}
\title{Validate Pushover app token values}
\usage{
check_valid_app(x, ...)

test_valid_app(x, ...)
}
\arguments{
\item{x}{Value to check}

\item{...}{Additional arguments passed to \code{\link[checkmate:checkString]{checkmate::check_string()}}}
}
\value{
\code{check_valid_app()} returns \code{TRUE} if the given value is a valid
Pushover app token or a string containing an error message otherwise

\code{test_valid_app()} returns a logical value indicating whether the
given value is a valid Pushover app token
}
\description{
\code{check_valid_app()} determines whether or not a given
application token is valid or not according to Pushover's specifications. It
does not determine whether or not the given token is associated with an
application.
}
\examples{
check_valid_app("azGDORePK8gMaC0QOYAMyEEuzJnyUi")
test_valid_app("azGDORePK8gMaC0QOYAMyEEuzJnyUi")
}
