% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valid_device.R
\name{check_valid_device}
\alias{check_valid_device}
\alias{test_valid_device}
\title{Validate Pushover device name values}
\usage{
check_valid_device(x, ...)

test_valid_device(x, ...)
}
\arguments{
\item{x}{Value to check}

\item{...}{Additional arguments passed to \code{\link[checkmate:checkString]{checkmate::check_string()}}}
}
\value{
\code{check_valid_device()} returns \code{TRUE} if the given value is a valid
Pushover device name or a string containing an error message otherwise

\code{test_valid_device()} returns a logical value indicating whether the
given value is a valid Pushover device name.
}
\description{
\code{check_valid_device()} determines whether a device name is
valid. Valid device names are strings up to 25 characters long and can
include letters, numbers, _, and -. It does not determine whether that device
is actually registered (see: \code{\link[=is.registered_device]{is.registered_device()}}.
}
\examples{
check_valid_device("phone")
test_valid_device("tablet")
}
