% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valid_receipt.R
\name{check_valid_receipt}
\alias{check_valid_receipt}
\alias{test_valid_receipt}
\title{Validate Pushover receipt ID values}
\usage{
check_valid_receipt(x, ...)

test_valid_receipt(x, ...)
}
\arguments{
\item{x}{Value to check}

\item{...}{Additional arguments passed to \code{\link[checkmate:checkString]{checkmate::check_string()}}}
}
\value{
\code{check_valid_receipt()} returns \code{TRUE} if the given value is a valid
Pushover receipt ID or a string containing an error message otherwise.

\code{test_valid_receipt()} returns a logical value indicating whether the
given value is a valid Pushover receipt ID.
}
\description{
\code{check_valid_receipt()} determines whether or not a given
message receipt is valid or not according to Pushover's specifications. It
does not determine whether or not the given receipt actually exists. Receipts
are 30-character strings containing letters and numbers ([A-Za-z0-9]).
}
\examples{
check_valid_receipt("KAWXTswy4cekx6vZbHBKbCKk1c1fdf")
test_valid_receipt("KAWXTswy4cekx6vZbHBKbCKk1c1fdf")
}
