% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valid_sound.R
\docType{data}
\name{check_valid_sound}
\alias{check_valid_sound}
\alias{test_valid_sound}
\alias{pushover_sounds}
\title{Validate Pushover sound values}
\format{
\code{pushover_sounds} is an object of class character of length 22 (i.e.,
a list of strings)
}
\usage{
check_valid_sound(x, ...)

test_valid_sound(x, ...)

pushover_sounds
}
\arguments{
\item{x}{Value to check}

\item{...}{Additional arguments passed to \code{\link[checkmate:checkChoice]{checkmate::check_choice()}}}
}
\value{
\code{check_valid_sound()} returns \code{TRUE} if the given value is a valid
Pushover sound or a string containing an error message otherwise.

\code{test_valid_sound()} returns a logical value indicating whether the
given value is a valid Pushover sound.
}
\description{
\code{check_valid_sound()} checks whether or not a given sound is
supported by Pushover. See \code{pushover_sounds} or
\url{https://pushover.net/api#sounds} for a complete list.

\code{pushover_sounds} is a list of sound names supported by Pushover
}
\examples{
check_valid_sound("cosmic")
test_valid_sound("cashregister")
}
\keyword{datasets}
