% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valid_user.R
\name{check_valid_user}
\alias{check_valid_user}
\alias{check_valid_group}
\alias{test_valid_user}
\alias{test_valid_group}
\title{Validate Pushover user and group values}
\usage{
check_valid_user(x, ...)

check_valid_group(x, ...)

test_valid_user(x, ...)

test_valid_group(x, ...)
}
\arguments{
\item{x}{Value to check}

\item{...}{Additional arguments passed to \code{\link[checkmate:checkString]{checkmate::check_string()}}}
}
\value{
\code{check_valid_user()} returns \code{TRUE} if the given value is a valid
Pushover receipt ID or a string containing an error message otherwise.

\code{test_valid_user()} and \code{test_valid_group()} return a logical value
indicating whether the given value is a valid user/group key.
}
\description{
\code{check_valid_user()} and \code{check_valid_group()} determine whether
or not a given user/group key is valid. They do not determine whether or not
the key is registered with Pushover.
}
\examples{
check_valid_user("uQiRzpo4DXghDmr9QzzfQu27cmVRsG")
test_valid_user("uQiRzpo4DXghDmr9QzzfQu27cmVRsG")
}
\seealso{
\code{\link[=verify_user]{verify_user()}} and \code{\link[=verify_group]{verify_group()}} determine whether or not
the given user/group key is registered with Pushover
}
