% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pushover_sounds.R
\name{get_pushover_sounds}
\alias{get_pushover_sounds}
\title{Get a list of sounds available for Pushover notifications}
\usage{
get_pushover_sounds(app = get_pushover_app())
}
\arguments{
\item{app}{application token (see \code{\link[=set_pushover_app]{set_pushover_app()}})}
}
\value{
A list of available sounds and their descriptions.
}
\description{
Get a list of sounds available for Pushover notifications
}
\examples{
\dontrun{
get_pushover_sounds(app = "azGDORePK8gMaC0QOYAMyEEuzJnyUi")
}
}
