% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpushover.R
\name{ggpushover}
\alias{ggpushover}
\title{Send a message with plot as picture attachment using Pushover}
\usage{
ggpushover(
  message,
  title = NULL,
  priority = 0,
  user = get_pushover_user(),
  app = get_pushover_app(),
  plot = ggplot2::last_plot(),
  device = NULL,
  sound = NULL,
  url = NULL,
  url_title = NULL,
  retry = 60,
  expire = 3600,
  callback = NULL,
  timestamp = NULL,
  scale = 1,
  width = graphics::par("din")[1],
  height = graphics::par("din")[2],
  units = c("in", "cm", "mm"),
  dpi = 300,
  limitsize = TRUE,
  file = "ggplot",
  ...
)
}
\arguments{
\item{message}{The message to be sent (max. 1024 characters)}

\item{title}{(optional) The message's title}

\item{priority}{Message priority (-2: silent, -1: quiet, 0: normal (default), 1: high, 2: emergency)}

\item{user}{user/group key (see \code{\link{set_pushover_user}})}

\item{app}{application token (see \code{\link{set_pushover_app}})}

\item{plot}{name of ggplot object to plot (default: last_plot())}

\item{device}{(optional) name of the device(s) to send message to. Defaults to all devices.}

\item{sound}{(optional) name of the sound to play (see \url{https://pushover.net/api#sounds})}

\item{url}{(optional) supplementary URL to display with message}

\item{url_title}{(optional) title to show for supplementary URL}

\item{retry}{(optional) how often (in seconds) to repeat emergency priority messages (min: 30 seconds; default: 60 seconds)}

\item{expire}{(optional) how long (in seconds) emergency priority messages will be retried (max: 86400 seconds; default: 3600 seconds)}

\item{callback}{(optional) callback URL to be visited (HTTP POST) once an emergency priority message has been acknowledged (\href{https://pushover.net/api#receipt}{details})}

\item{timestamp}{(optional) a Unix timestamp containing the date and time to display to the user instead of the time at which the message was received}

\item{scale}{scaling factor}

\item{width}{width (defaults to the width of current plotting window)}

\item{height}{height (defaults to the height of current plotting window)}

\item{units}{units for width and height when either one is explicitly specified
(in, cm, or mm)}

\item{dpi}{dpi to use for raster graphics}

\item{limitsize}{when TRUE (the default), ggsave will not save images larger
than 50x50 inches, to prevent the common error of specifying dimensions in pixels.}

\item{file}{prefix for filenames (defaults to \code{ggplot})}

\item{...}{other arguments passed to graphics device}
}
\value{
an invisible list containing the following fields:
\itemize{
    \item \code{status}: request status (1 = success)
    \item \code{request}: unique request ID
    \item \code{raw}: the raw \code{\link[httr]{response}} object
    \item \code{receipt}: a receipt ID (only for emergency priority messages)
    \item \code{errors}: a list of error messages (only for unsuccessful requests)
}
}
\description{
\code{ggpushover} sends a message (push notification) to a user or group with ggplot object as picture attachment.
Messages can be given different priorities, play different sounds, or require
acknowledgments. The \code{pushover_normal}, \code{pushover_silent},
\code{pushover_quiet}, \code{pushover_high}, and \code{pushover_emergency}
functions send messages with those priorities.
}
\examples{
\dontrun{
ggpushover(message = "Hola Mundo!", plot = last_plot())
}
}
