% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_rename.R
\name{group_rename}
\alias{group_rename}
\title{Rename a delivery group}
\usage{
group_rename(group, name, app = get_pushover_app())
}
\arguments{
\item{group}{group key}

\item{name}{new group name}

\item{app}{application token (see \code{\link[=set_pushover_app]{set_pushover_app()}})}
}
\value{
An invisible list containing the following fields:
\itemize{
\item \code{status}: request status (\code{1} = success)
\item \code{request}: unique request ID
\item \code{raw}: the raw \link[httr:response]{httr::response} object
}
}
\description{
Rename a delivery group
}
\examples{
\dontrun{
group_rename(
  group = "gznej3rKEVAvPUxu9vvNnqpmZpokzF",
  name = "Coffee Party"
)
}
}
