% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{headers,PushoverResponse-method}
\alias{headers}
\alias{headers,PushoverResponse-method}
\title{Get the HTTP headers returned by a Pushover API query}
\usage{
\S4method{headers}{PushoverResponse}(object)
}
\arguments{
\item{object}{A \code{\link{PushoverResponse-class}} object}
}
\value{
A list of values from the HTTP header
}
\description{
Aside from standard HTTP parameters, headers in Pushover API responses also
contain information about the application, including the app's limit on the
number of messages sent per month ('\code{x-limit-app-limit}'), how many
messages remain ('\code{x-limit-app-remaining}'), and when the message
counter will be reset ('\code{x-limit-app-reset}').
}

