% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.valid_device.R
\name{is.valid_device}
\alias{is.valid_device}
\title{Determine whether or not a given device name is valid}
\usage{
is.valid_device(device)
}
\arguments{
\item{device}{one or more device names (e.g., "phone", c("phone", "tablet"))}
}
\value{
Logical value(s) indicating whether the corresponding device name is
valid (\code{TRUE}) or not (\code{FALSE})
}
\description{
Valid device names are strings up to 25 characters long and can include
letters, numbers, _, and -.
}
\note{
\code{is.valid_device} only determines whether a device name is valid.
It does not determine whether that device is actually registered (see:
\code{\link{is.device}}).
}
\examples{
is.valid_device("my_phone")

}

