% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.valid_user.R
\name{is.valid_user}
\alias{is.valid_user}
\alias{is.valid_group}
\title{Determine whether a user or group key is valid}
\usage{
is.valid_user(user)

is.valid_group(group)
}
\arguments{
\item{user}{user/group key to verify (e.g., \code{"uQiRzpo4DXghDmr9QzzfQu27cmVRsG"})}

\item{group}{user/group key to verify (e.g., \code{"gznej3rKEVAvPUxu9vvNnqpmZpokzF"})}
}
\value{
A logical value indicating whether or not the given user/group key is
valid.
}
\description{
\code{is.valid_user()} and \code{is.valid_group()} determine whether
or not a given user/group key is valid. They do not determine whether or not
the key is registered with Pushover.
}
\examples{
\dontrun{
is.valid_user("uQiRzpo4DXghDmr9QzzfQu27cmVRsG")
}
}
\seealso{
\code{\link[=verify_user]{verify_user()}} and \code{\link[=verify_group]{verify_group()}} determine whether or not
the given user key is registered with Pushover
}
