% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushover_sounds.R
\docType{data}
\name{pushover_sounds}
\alias{pushover_sounds}
\title{Sounds for push notifications}
\format{
\code{pushover_sounds} is an object of class character of length 22 (i.e.,
a list of strings)
}
\usage{
pushover_sounds
}
\description{
\code{pushover_sounds} is a list of sound names supported by Pushover
}
\keyword{datasets}
