% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushoverr-package.R
\docType{package}
\name{pushoverr-package}
\alias{pushoverr}
\alias{pushoverr-package}
\title{pushoverr: Send Push Notifications using Pushover}
\description{
Send push notifications to mobile devices or the desktop using Pushover. These notifications can display job status, results, scraped web data, or any other text or numeric data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/briandconnelly/pushoverr}
  \item Report bugs at \url{https://github.com/briandconnelly/pushoverr/issues}
}

}
\author{
\strong{Maintainer}: Brian Connelly \email{bdc@bconnelly.net} (\href{https://orcid.org/0000-0002-9948-0379}{ORCID})

}
\keyword{internal}
