% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushoverr.R
\docType{package}
\name{pushoverr}
\alias{pushoverr}
\alias{pushoverr-package}
\title{pushoverr: Send push notifications using Pushover}
\description{
pushoverr is package for sending push notifications to mobile devices
(iOS and Android) and the desktop using
\href{https://pushover.net}{Pushover}. Messages can be quickly sent using the
\code{\link{pushover}} function. Apple Watch users can push small bits of
data to their watches using \code{\link{update_glance}}.
}
\details{
To use pushoverr, you will first need your \emph{user key} and an
\emph{application token}. The user key can be found within the settings of
the app (e.g., \code{uQiRzpo4DXghDmr9QzzfQu27cmVRsG}) or by logging in to
\url{https://pushover.net}. To receive an application token, register a new
application after logging in to \url{https://pushover.net/apps}.

For more information, browse the package index by following the link
at the bottom of this help page. For release information, run
\code{news(package = "pushoverr")}.
}
\section{Disclaimer}{
 This package and its author are not affiliated with
\href{https://superblock.net}{Superblock, LLC}, creators of Pushover.
}
\author{
Brian Connelly
}
\seealso{
GitHub Repository: \url{https://github.com/briandconnelly/pushoverr}

Bug Reports and Feature Requests: \url{https://github.com/briandconnelly/pushoverr/issues}

Pushover: \url{https://pushover.net}

Pushover API: \url{https://pushover.net/api}
}

