% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_for_pushover_status.R
\name{stop_for_pushover_status}
\alias{stop_for_pushover_status}
\title{Raise an Error if Pushover Request Failed}
\usage{
stop_for_pushover_status(x)
}
\arguments{
\item{x}{a \code{\link[httr]{response}} object returned by an API call}
}
\value{
If request was successful, the response (invisibly).
}
\description{
Examines a response from a Pushover API call for errors. If there were
errors, execution is stopped, and an error message is shown.
}

