% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_glance.R
\name{update_glance}
\alias{update_glance}
\title{Update a Pushover glance data}
\usage{
update_glance(user = get_pushover_user(), app = get_pushover_app(),
  device = NULL, ...)
}
\arguments{
\item{user}{user/group key (see \code{\link{set_pushover_user}})}

\item{app}{application token (see \code{\link{set_pushover_app}})}

\item{device}{(optional) name of the device(s) to send message to. Defaults
to all devices.}

\item{...}{Values for glance data fields. Must provide at least one of:
\itemize{
    \item \code{title}: a description of the data being shown, such as "Widgets Sold" (max. 100 characters)
    \item \code{text}: the main line of data, used on most screens (max. 100 characters)
    \item \code{subtext}: a second line of data (max. 100 characters)
    \item \code{count}: integer value shown on smaller screens; useful for simple counts
    \item \code{percent}: integer percent value (0..100) shown on some screens as a progress bar/circle
}}
}
\value{
an invisible list containing the following fields:
\itemize{
    \item \code{status}: request status (1 = success)
    \item \code{request}: unique request ID
    \item \code{raw}: the raw \code{\link[httr]{response}} object
    \item \code{errors}: a list of error messages (only for unsuccessful requests)
}
}
\description{
Glances allow you to push small pieces of data to a frequently-updated screen
such as a smartwatch or a lock screen.
}
\note{
Glances are currently in beta, and features may change.
}
\examples{
\dontrun{
update_glance(count = 37)
}
}

