% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_user.R
\name{validate_user}
\alias{is.valid_group}
\alias{is.valid_key}
\alias{is.valid_user}
\alias{validate_group}
\alias{validate_key}
\alias{validate_user}
\title{Determine whether or not a given user/group key is valid}
\usage{
validate_user(user, app = get_pushover_app(), device = NULL)

validate_group(user, app = get_pushover_app(), device = NULL)

validate_key(user, device = NA_character_, ...)

is.valid_user(user, app = get_pushover_app(), device = NULL)

is.valid_group(user, app = get_pushover_app(), device = NULL)

is.valid_key(user, device = NA, ...)
}
\arguments{
\item{user}{user/group key to validate}

\item{app}{application token (see \code{\link{set_pushover_app}})}

\item{device}{(optional) device to validate. If supplied the device must be registered to the given user's account.}
}
\value{
a list containing the following fields:
\itemize{
    \item \code{status}: request status (1 = success)
    \item \code{devices}: a list of the user's devices
    \item \code{request}: unique request ID
    \item \code{errors}: a list of error messages (only for unsuccessful requests)
    \item \code{raw}: the raw \code{\link[httr]{response}} object
}
}
\description{
\code{validate_user} determines whether or not the given user or
group is valid (i.e., a registered user or group).

\code{validate_group} is an alias for \code{validate_user}

\code{is.valid_key} is deprecated in favor of
\code{is.valid_user}

\code{is.valid_user} indicates whether or not a given user ID
is registered with Pushover

\code{is.valid_group} is an alias for \code{is.valid_user}

\code{is.valid_key} is deprecated in favor of \code{is.valid_user} or \code{is.valid_group}
}
\examples{
validate_user(user = "uQiRzpo4DXghDmr9QzzfQu27cmVRsG")

}

