% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_pushover_sounds}
\alias{get_pushover_sounds}
\title{Get a list of available message sounds}
\usage{
get_pushover_sounds(...)
}
\arguments{
\item{...}{An application token can be specified with \code{token}}
}
\value{
A list of available sounds and their descriptions
}
\description{
\code{get_pushover_sounds} queries Pushover to receive a list of the sounds
that may be played on a user's device when a message is received. Sounds are
specified when the message is sent.
}
\examples{
\dontrun{
sounds <- get_pushover_sounds()
for (s in names(sounds))
{
    cat(paste(s, ":", sounds[s], "\\n"))
}
}
}

