% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpushover.R
\name{ggpushover}
\alias{ggpushover}
\title{Send a message with an attached plot}
\usage{
ggpushover(plot = ggplot2::last_plot(), message, ...)
}
\arguments{
\item{plot}{Plot to save, defaults to last plot displayed}

\item{message}{The message to be sent (max. 1024 characters). Messages use
\code{\link[glue:glue]{glue::glue()}} for formatting and interpolation.}

\item{...}{
  Arguments passed on to \code{\link[=pushover]{pushover}}
  \describe{
    \item{\code{title}}{(optional) The message's title. Titles use \code{\link[glue:glue]{glue::glue()}} for
formatting and interpolation.}
    \item{\code{priority}}{Message priority (\code{-2}: silent, \code{-1}: quiet, \code{0}: normal
(default), \code{1}: high, \code{2}: emergency)}
    \item{\code{attachment}}{Path of file attachment to include. File must be image
format (bmp, jpg, png, tif) and no larger than 2.6 MB.}
    \item{\code{user}}{user/group key (see \code{\link[=set_pushover_user]{set_pushover_user()}})}
    \item{\code{app}}{application token (see \code{\link[=set_pushover_app]{set_pushover_app()}})}
    \item{\code{device}}{(optional) name of the device(s) to send message to. Defaults
to all devices.}
    \item{\code{sound}}{(optional) name of the sound to play (see
\url{https://pushover.net/api#sounds})}
    \item{\code{url}}{(optional) supplementary URL to display with message}
    \item{\code{url_title}}{(optional) title to show for supplementary URL}
    \item{\code{format}}{Message formatting. If \code{html} (default), messages can include a
\href{https://pushover.net/api#html}{limited subset} of HTML formatting. If
\code{monospace}, text is formatted using monospace font.}
    \item{\code{retry}}{(optional) how often (in seconds) to repeat emergency priority
messages (min: 30 seconds; default: 60 seconds)}
    \item{\code{expire}}{(optional) how long (in seconds) emergency priority messages
will be retried (max: 86400 seconds; default: 3600 seconds)}
    \item{\code{callback}}{(optional) callback URL to be visited (HTTP POST) once an
emergency priority message has been acknowledged
(\href{https://pushover.net/api#receipt}{details})}
    \item{\code{timestamp}}{(optional) a Unix timestamp containing the date and time to
display to the user instead of the time at which the message was received}
  }}
}
\value{
an invisible list containing the following fields:
\itemize{
\item \code{status}: request status (1 = success)
\item \code{request}: unique request ID
\item \code{raw}: the raw \link[httr:response]{httr::response} object
\item \code{receipt}: a receipt ID (only for emergency priority messages)
\item \code{errors}: a list of error messages (only for unsuccessful requests)
}
}
\description{
\code{ggpushover()} sends a \link[ggplot2]{ggplot2} plot and message to your
devices as a push notification
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(wt, mpg)) +
  geom_point()
ggpushover(p, message = "We're gonna need to shed some weight!")
}
