% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.valid_user.R
\name{is.valid_user}
\alias{is.valid_group}
\alias{is.valid_user}
\title{Determine whether a user or group key is valid}
\usage{
is.valid_user(user)

is.valid_group(group)
}
\arguments{
\item{user}{user/group key to verify (e.g., "uQiRzpo4DXghDmr9QzzfQu27cmVRsG")}

\item{group}{user/group key to verify (e.g., "gznej3rKEVAvPUxu9vvNnqpmZpokzF")}
}
\value{
A logical value indicating whether (\code{TRUE}) or not
(\code{FALSE}) the given user/group key is valid.
}
\description{
\code{is.valid_user} and \code{is.valid_group} determine whether
or not a given user/group key is valid. They do not determine whether or not
the key is registered with Pushover.

\code{is.valid_group} is an alias for \code{is.valid_user}
}
\examples{
\dontrun{
is.valid_user("uQiRzpo4DXghDmr9QzzfQu27cmVRsG")
}
}
\seealso{
\code{\link{verify_user}} and \code{\link{verify_group}} determine
whether or not the given user key is registered with Pushover
}

