\docType{data}
\name{pushover_sounds}
\alias{pushover_sounds}
\title{\code{pushover_sounds} is a vector containing all of the sounds that Pushover
supports}
\format{\preformatted{ chr [1:22] "bike" "bugle" "cashregister" "classical" "cosmic" "falling" "gamelan" "incoming" "intermission" "magic" "mechanical" ...
}}
\usage{
pushover_sounds
}
\description{
\code{pushover_sounds} is a vector containing all of the
sounds that Pushover supports
}
\keyword{datasets}

