% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is.valid_sound}
\alias{is.valid_sound}
\title{Validate Pushover values}
\usage{
is.valid_sound(sound)
}
\arguments{
\item{sound}{a string containing a sound name}
}
\value{
\code{is.valid_sound()} returns a logical value indicating whether or not
the given sound is supported by Pushover.
}
\description{
\code{is.valid_sound()} checks whether or not a given sound is
supported by Pushover. See \url{https://pushover.net/api#sounds} for a complete
list.
}
\examples{
is.valid_sound("cosmic")
}
