% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is.valid_receipt}
\alias{is.valid_receipt}
\alias{is.valid_user}
\alias{is.valid_group}
\alias{is.valid_sound}
\title{Validate Pushover values}
\usage{
is.valid_receipt(receipt)

is.valid_user(user)

is.valid_group(group)

is.valid_sound(sound)
}
\arguments{
\item{receipt}{A message receipt (e.g., \code{"KAWXTswy4cekx6vZbHBKbCKk1c1fdf"})}

\item{user}{user/group key to verify (e.g., \code{"uQiRzpo4DXghDmr9QzzfQu27cmVRsG"})}

\item{group}{user/group key to verify (e.g., \code{"gznej3rKEVAvPUxu9vvNnqpmZpokzF"})}

\item{sound}{a string containing a sound name}
}
\value{
\code{is.valid_receipt()} returns a boolean value for each given message
receipt ID indicating whether or not that receipt ID is valid.

A logical value indicating whether or not the given user/group key is
valid.

\code{is.valid_sound()} returns a logical value indicating whether or not
the given sound is supported by Pushover.
}
\description{
\code{is.valid_receipt()} determines whether or not a given message
receipt is valid or not according to Pushover's specifications. It does not
determine whether or not the given receipt actually exists. Receipts are
30-character strings containing letters and numbers ([A-Za-z0-9]).

\code{is.valid_user()} and \code{is.valid_group()} determine whether
or not a given user/group key is valid. They do not determine whether or not
the key is registered with Pushover.

\code{is.valid_sound()} checks whether or not a given sound is
supported by Pushover. See \url{https://pushover.net/api#sounds} for a complete
list.
}
\examples{
\dontrun{
is.valid_receipt(receipt = "KAWXTswy4cekx6vZbHBKbCKk1c1fdf")
}

\dontrun{
is.valid_user("uQiRzpo4DXghDmr9QzzfQu27cmVRsG")
}
is.valid_sound("cosmic")
}
\seealso{
\code{\link[=verify_user]{verify_user()}} and \code{\link[=verify_group]{verify_group()}} determine whether or not
the given user key is registered with Pushover
}
